requirements.txt
annotated-doc==0.0.4
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.1.2
boolean.py==5.0
CacheControl==0.14.4
cachetools==5.3.3
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==44.0.1
cyclonedx-python-lib==9.1.0
defusedxml==0.7.1
ecdsa==0.19.1
fastapi==0.122.0
filelock==3.20.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
greenlet==3.0.3
gspread==6.1.0
h11==0.16.0
httptools==0.6.1
idna==3.7
license-expression==30.4.4
markdown-it-py==4.0.0
mdurl==0.1.2
msgpack==1.1.2
oauthlib==3.2.2
packageurl-python==0.17.6
packaging==25.0
passlib==1.7.4
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.9.0
platformdirs==4.5.0
py-serializable==2.1.0
pyasn1==0.4.8
pyasn1_modules==0.4.0
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
Pygments==2.19.2
pyparsing==3.2.5
python-dotenv==1.0.1
python-http-client==3.3.7
python-jose==3.4.0
python-multipart==0.0.18
pytz==2025.2
PyYAML==6.0.1
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.2.0
rsa==4.9
sendgrid==6.11.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
starkbank-ecdsa==2.2.0
starlette==0.50.0
StrEnum==0.4.15
toml==0.10.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.27.1
watchfiles==0.21.0
websockets==12.0
